# specs/001-enterprise-chat-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: InternalChat API
  version: 1.0.0
  description: >
    HTTP contract for the enterprise internal chat platform.
    All endpoints require a Keycloak-issued bearer token and a resource-scoped
    authorization policy. Errors use Problem Details (RFC 9457).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /me/sessions/{sessionId}:
    delete:
      tags: [Directory]
      summary: Revoke one of this employee's own sessions (FR-005)
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /conversations:
    get:
      tags: [Conversations]
      summary: List conversations this employee belongs to, with unread counts
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 100 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversationPage" }
    post:
      tags: [Conversations]
      summary: Create a direct or group conversation (FR-007, FR-008)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateConversationRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "200":
          description: >
            An existing direct conversation between this pair was returned rather than
            creating a duplicate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "422":
          description: A named member is deactivated and cannot be added
  /conversations/{conversationId}:
    parameters:
      - { name: conversationId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Conversations]
      summary: Conversation detail
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: >
        Not a member of this resource. Indistinguishable from 404 in body and timing, so that
        a refusal never reveals that the resource exists (SC-017).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457. Never contains stack traces, SQL, or internal hostnames.
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
...
    Conversation:
      type: object
      required: [id, kind, historyVisibility, lastSeq]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [direct, group] }
        name: { type: string, nullable: true }
        historyVisibility:
          type: string
          enum: [from_join, full]
          description: Displayed to members, per US3 scenario 4.
        lastSeq: { type: integer, format: int64 }
        unreadCount: { type: integer }
        mutedUntil: { type: string, format: date-time, nullable: true }
        memberCount: { type: integer }
        lastMessage: { $ref: "#/components/schemas/Message" }
    ConversationPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Conversation" }
        nextCursor: { type: string, nullable: true }
    CreateConversationRequest:
      type: object
      required: [kind, memberIds]
      properties:
        kind: { type: string, enum: [direct, group] }
        name: { type: string, nullable: true, maxLength: 200 }
        memberIds:
          type: array
          items: { type: string, format: uuid }
          minItems: 1
        historyVisibility: { type: string, enum: [from_join, full], default: from_join }
...
    Message:
      type: object
      required: [id, conversationId, seq, authorId, sentAt]
      properties:
        id: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        seq:
          type: integer
          format: int64
          description: Server-assigned, gapless per conversation. The ordering authority.
        authorId: { type: string, format: uuid }
        clientMessageKey: { type: string }
        body: { type: string, nullable: true, description: Null when deleted }
        sentAt: { type: string, format: date-time }
        editedAt: { type: string, format: date-time, nullable: true }
        deletedAt: { type: string, format: date-time, nullable: true }
        mentions:
          type: array
          items: { type: string, format: uuid }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
...
    Attachment:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [image, video] }
        contentType: { type: string }
        byteSize: { type: integer, format: int64 }
        durationSeconds: { type: integer, nullable: true }
        fileName: { type: string }
        scanStatus: { type: string, enum: [pending, clean, infected, failed] }
        contentUrl:
          type: string
          nullable: true
          description: Present only when scanStatus is clean. Always authorized per request.
        posterUrl: { type: string, nullable: true }
...